requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
uvicorn==0.22.0
pydantic==1.10.7
docker==6.1.2
python-multipart==0.0.6
requests==2.28.2
